Getting the wiring in place
