+++ features/src/main/resources/features.xml
@@ -1,6 +1,8 @@
 <features name="ignite-poc" xmlns="http://karaf.apache.org/xmlns/features/v1.2.0">
 
 <!--    <repository>mvn:org.apache.ignite/ignite-osgi-karaf/2.12.0/xml/features</repository>-->
+    <repository>mvn:org.apache.aries.jax.rs/org.apache.aries.jax.rs.features/${aries.jax.rs.whiteboard.version}/xml</repository>
+    <repository>mvn:org.apache.cxf.karaf/apache-cxf/${cxf.version}/xml/features</repository>
 
     <feature name="ignite-poc" version="${project.version}">
         <feature>guava</feature>
@@ -17,8 +19,25 @@
     </feature>
 
     <feature name="ignite-core" version="${project.version}">
+        <feature>ignite-poc-rest</feature>
         <bundle dependency="true">mvn:org.apache.servicemix.bundles/org.apache.servicemix.bundles.javax-cache-api/1.0.0_1</bundle>
         <bundle dependency="true">wrap:mvn:org.apache.ignite/ignite-osgi/2.12.0</bundle>
         <bundle>wrap:mvn:org.apache.ignite/ignite-core/2.12.0</bundle>
+    </feature>
+
+    <feature name="ignite-poc-rest" version="${project.version}">
+
+        <feature>cxf</feature>
+        <feature>cxf-jaxrs</feature>
+        <feature>cxf-sse</feature>
+        <feature>cxf-rs-description-openapi-v3</feature>
+        <feature>cxf-rs-security-oauth2</feature>
+
+        <bundle>mvn:org.apache.aries.component-dsl/org.apache.aries.component-dsl.component-dsl/1.2.2</bundle>
+        <bundle>mvn:org.apache.aries.jax.rs/org.apache.aries.jax.rs.openapi.resource/${aries.jax.rs.whiteboard.version}</bundle>
+        <feature>aries-jax-rs-whiteboard</feature>
+        <feature>aries-jax-rs-whiteboard-jackson</feature>
+        <bundle>mvn:javax.annotation/javax.annotation-api/1.3.2</bundle>
+
     </feature>
 </features>
