+++ ignite-worker/src/main/java/org/opennms/poc/ignite/worker/ignite/IgnitionFactory.java
@@ -1,14 +1,12 @@
 package org.opennms.poc.ignite.worker.ignite;
 
+import lombok.AllArgsConstructor;
 import org.apache.ignite.Ignite;
 import org.apache.ignite.Ignition;
-import org.apache.ignite.configuration.IgniteConfiguration;
 
 public class IgnitionFactory {
-
-    public static Ignite create(IgniteConfiguration igniteConfiguration) {
-        return Ignition.start(igniteConfiguration);
-//        return Ignition.start(new IgniteConfiguration());
+    public static Ignite create(WorkerIgniteConfiguration workerIgniteConfiguration) {
+        return Ignition.start(workerIgniteConfiguration.prepareIgniteConfiguration());
     }
 
 }
