+++ ignite-worker/src/main/java/org/opennms/poc/ignite/worker/ignite/service/AllRepeatedService.java
@@ -1,15 +1,19 @@
 package org.opennms.poc.ignite.worker.ignite.service;
 
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.ignite.Ignite;
 import org.apache.ignite.resources.IgniteInstanceResource;
 import org.apache.ignite.services.Service;
 
 import java.util.UUID;
 
+@Slf4j
+//@RequiredArgsConstructor
 public class AllRepeatedService implements Service {
     private static final long serialVersionUID = 0L;
-    private final UUID myNodeId;
 
+    private final UUID myNodeId;
     @IgniteInstanceResource
     private Ignite ignite;
