+++ pom.xml
@@ -25,6 +25,8 @@
 
         <spring-boot.version>2.6.4</spring-boot.version>
         <slf4j.version>1.7.33</slf4j.version>
+        <aries.jax.rs.whiteboard.version>2.0.1</aries.jax.rs.whiteboard.version>
+        <cxf.version>3.5.0</cxf.version>
     </properties>
 
     <modules>
